feat: add proxy chaining template
This template is meant to be used with multiple
input OpenAPI specs. Each of the input specs
represent a proxy that we want to chain to.

The idea is that we can have a common/parent proxy
with a base path that is shared among the children,
but still allow the children proxy to declare their
own unique operations under this base common basepath.